EN-51742: Add join support in rollup
Move functions from soql—server-pg to common-pg so that they can be shared between rollup and query.
