+++ common-pg/src/main/scala/com/socrata/pg/query/ExplainInfo.scala
@@ -1,0 +1,3 @@
+package com.socrata.pg.query
+
+case class ExplainInfo(query: String, explainPlan: String)
